+++ addon/templates/components/head-layout.hbs
@@ -1,0 +1,1 @@
+<meta name="ember-cli-head-start" />{{head-content}}<meta name="ember-cli-head-end" />
