+++ app/components/head-layout.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import layout from 'ember-cli-head/templates/components/head-layout';
+
+export default Ember.Component.extend({
+  tagName: '',
+  layout
+});
