--- app/instance-initializers/head.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-cli-head/instance-initializers/head';
